+++ src/main/java/com/mosaicchurchaustin/oms/data/response/FeatureConfigResponse.java
@@ -10,11 +10,13 @@ import lombok.Getter;
 @Builder
 public class FeatureConfigResponse {
     private boolean groupMeEnabled;
+    private boolean ordersOpen;
     private OrderStatus printOnTransitionToStatus;
 
     public static FeatureConfigResponse from(final FeatureConfigEntity featureConfigEntity) {
         return FeatureConfigResponse.builder()
                 .groupMeEnabled(featureConfigEntity.isGroupMeEnabled())
+                .ordersOpen(featureConfigEntity.isOrdersOpen())
                 .printOnTransitionToStatus(featureConfigEntity.getPrintOnTransitionToStatus())
                 .build();