+++ src/main/resources/db/migration/V00033__add-feature-config.sql
@@ -1,0 +1,7 @@
+alter table features add column
+(
+    orders_open bit          not null default true
+);
+
+
+