+++ src/main/resources/changelog.md
@@ -5,6 +5,7 @@
 ### Added
 - Finished label printing with QR code scanning in dashboards
 - Fallback page when service has not yet started up
+- Toggle to enable/disable order acceptance
 
 ### Fixed
 - Navigation bar issues